refactor: use singular form for file name
