+++ frontend/src/components/matchEvents/TimelineEvent.tsx
@@ -1,4 +1,4 @@
-import type { Match } from '@/types/matches';
+import type { Match } from '@/types/match';
 import type { MatchEvent } from '@/types/matchEvents';
 
 import EventCard from './EventCard';
