+++ frontend/src/components/matches/MatchDayAccordion.test.tsx
@@ -1,7 +1,7 @@
 import { cleanup, render, screen, within } from '@testing-library/react';
 import { afterEach, describe, expect, it, vi } from 'vitest';
 
-import type { Match } from '@/types/matches';
+import type { Match } from '@/types/match';
 
 import MatchCard from './MatchCard';
 import MatchDayAccordion from './MatchDayAccordion';
