+++ frontend/src/utils/matchEvents/matchHeaderHelper.ts
@@ -1,4 +1,4 @@
-import type { Match } from '@/types/matches';
+import type { Match } from '@/types/match';
 import type { MatchEvent } from '@/types/matchEvents';
 
 export function formatMatchDate(kickoffTime: string) {
